9043/refactor: methods transferred from UBSClientService to other services
