+++ core/src/test/java/greencity/controller/OrderControllerTest.java
@@ -32,12 +32,16 @@ import greencity.exception.handler.CustomExceptionHandler;
 import greencity.repository.OrderRepository;
 import greencity.repository.UBSUserRepository;
 import greencity.repository.UserRepository;
+import greencity.service.ubs.AddressService;
+import greencity.service.ubs.CertificateService;
+import greencity.service.ubs.EventService;
 import greencity.service.ubs.NotificationService;
 import greencity.service.ubs.UBSClientService;
 import greencity.service.ubs.UBSManagementService;
 import greencity.service.ubs.order.OrderCheckoutService;
 import greencity.service.ubs.order.OrderService;
 import greencity.service.ubs.payment.ProcessPaymentService;
+import greencity.service.ubs.tariff.TariffService;
 import greencity.service.ubs.wayforpay.WayForPayRedirectService;
 import greencity.service.ubs.wayforpay.WayForPayResultService;
 import java.security.Principal;
@@ -96,6 +100,18 @@ class OrderControllerTest {
     @Mock
     private OrderService orderService;
 
+    @Mock
+    private TariffService tariffService;
+
+    @Mock
+    private AddressService addressService;
+
+    @Mock
+    private EventService eventService;
+
+    @Mock
+    private CertificateService certificateService;
+
     @InjectMocks
     OrderController orderController;
 
@@ -150,7 +166,7 @@ class OrderControllerTest {
         mockMvc.perform(get(ubsLink + "/certificate/{code}", certificateCode)
                 .principal(principal))
             .andExpect(status().isOk());
-        verify(ubsClientService).checkCertificate(certificateCode, "35467585763t4sfgchjfuyetf");
+        verify(certificateService).checkCertificate(certificateCode, "35467585763t4sfgchjfuyetf");
     }
 
     @Test
@@ -270,7 +286,7 @@ class OrderControllerTest {
             .principal(principal))
             .andExpect(status().isOk());
 
-        verify(ubsClientService, times(1))
+        verify(eventService, times(1))
             .getAllEventsForOrder(1L, "test@gmail.com", "en");
     }
 
@@ -342,67 +358,67 @@ class OrderControllerTest {
     @Test
     void checkIfTariffExistsById_Returns200_WhenTariffExists() throws Exception {
         Long tariffId = 1L;
-        when(ubsClientService.checkIfTariffExistsById(tariffId)).thenReturn(true);
+        when(tariffService.checkIfTariffExistsById(tariffId)).thenReturn(true);
 
         mockMvc.perform(get(ubsLink + "/check-if-tariff-exists/{id}", tariffId)
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
             .andExpect(content().string(""));
 
-        verify(ubsClientService).checkIfTariffExistsById(tariffId);
+        verify(tariffService).checkIfTariffExistsById(tariffId);
     }
 
     @Test
     void checkIfTariffExistsById_Returns404_WhenTariffDoesNotExist() throws Exception {
         Long tariffId = 999999L;
-        when(ubsClientService.checkIfTariffExistsById(tariffId)).thenReturn(false);
+        when(tariffService.checkIfTariffExistsById(tariffId)).thenReturn(false);
 
         mockMvc.perform(get(ubsLink + "/check-if-tariff-exists/{id}", tariffId)
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isNotFound())
             .andExpect(content().string(""));
 
-        verify(ubsClientService).checkIfTariffExistsById(tariffId);
+        verify(tariffService).checkIfTariffExistsById(tariffId);
     }
 
     @Test
     void getAllLocationsTest() throws Exception {
         List<LocationsDto> locationsDtoList = Arrays.asList(new LocationsDto(), new LocationsDto());
-        when(ubsClientService.getAllLocations()).thenReturn(locationsDtoList);
+        when(addressService.getAllLocations()).thenReturn(locationsDtoList);
 
         mockMvc.perform(get(ubsLink + "/locations")
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
             .andExpect(content().json(new ObjectMapper().writeValueAsString(locationsDtoList)));
 
-        verify(ubsClientService).getAllLocations();
+        verify(addressService).getAllLocations();
     }
 
     @Test
     void getTariffIdByLocationIdTest() throws Exception {
         Long locationId = 1L;
         List<Long> tariffId = List.of(2L);
-        when(ubsClientService.getTariffIdByLocationId(locationId)).thenReturn(tariffId);
+        when(tariffService.getTariffIdByLocationId(locationId)).thenReturn(tariffId);
 
         mockMvc.perform(get(ubsLink + "/tariffs/{locationId}", locationId)
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
             .andExpect(content().string(String.valueOf(tariffId)));
 
-        verify(ubsClientService).getTariffIdByLocationId(locationId);
+        verify(tariffService).getTariffIdByLocationId(locationId);
     }
 
     @Test
     void getAllLocationsByCourierIdTest() throws Exception {
         Long id = 1L;
         List<LocationsDto> locationsDtoList = Arrays.asList(new LocationsDto(), new LocationsDto());
-        when(ubsClientService.getAllLocationsByCourierId(id)).thenReturn(locationsDtoList);
+        when(addressService.getAllLocationsByCourierId(id)).thenReturn(locationsDtoList);
 
         mockMvc.perform(get(ubsLink + "/locationsByCourier/" + id)
             .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
             .andExpect(content().json(new ObjectMapper().writeValueAsString(locationsDtoList)));
 
-        verify(ubsClientService).getAllLocationsByCourierId(id);
+        verify(addressService).getAllLocationsByCourierId(id);
     }
 }
