+++ service-api/src/main/java/greencity/service/ubs/CertificateService.java
@@ -1,5 +1,6 @@
 package greencity.service.ubs;
 
+import greencity.dto.certificate.CertificateDto;
 import greencity.dto.certificate.CertificateDtoForAdding;
 import greencity.dto.certificate.CertificateDtoForSearching;
 import greencity.dto.pageble.PageableDto;
@@ -29,4 +30,14 @@ public interface CertificateService {
      */
     PageableDto<CertificateDtoForSearching> getCertificatesWithFilter(CertificatePage certificatePage,
         CertificateFilterCriteria certificateFilterCriteria);
+
+    //TODO add test
+    /**
+     * Methods return status of entered certificate, empty string if absent.
+     *
+     * @param code {@link String} code of certificate.
+     * @return {@link CertificateDto} which contains status.
+     * @author Oleh Bilonizhka
+     */
+    CertificateDto checkCertificate(String code, String userUuid);
 }
