+++ service/src/main/java/greencity/service/ubs/CertificateServiceImpl.java
@@ -1,38 +1,21 @@
 package greencity.service.ubs;
 
 import static greencity.constant.ErrorMessage.CERTIFICATE_EXIST;
-import static greencity.constant.ErrorMessage.CERTIFICATE_EXPIRED;
-import static greencity.constant.ErrorMessage.CERTIFICATE_IS_NOT_ACTIVATED;
-import static greencity.constant.ErrorMessage.CERTIFICATE_IS_USED;
-import static greencity.constant.ErrorMessage.CERTIFICATE_NOT_FOUND;
 import static greencity.constant.ErrorMessage.CERTIFICATE_NOT_FOUND_BY_CODE;
-import static greencity.constant.ErrorMessage.SOME_CERTIFICATES_ARE_INVALID;
-import static greencity.constant.ErrorMessage.TOO_MANY_CERTIFICATES;
-import static java.util.stream.Collectors.joining;
-import greencity.constant.AppConstant;
 import greencity.constant.ErrorMessage;
 import greencity.dto.certificate.CertificateDto;
 import greencity.dto.certificate.CertificateDtoForAdding;
 import greencity.dto.certificate.CertificateDtoForSearching;
-import greencity.dto.order.OrderResponseDto;
-import greencity.dto.order.OrderWayForPayClientDto;
 import greencity.dto.pageble.PageableDto;
 import greencity.entity.order.Certificate;
-import greencity.entity.order.Order;
 import greencity.enums.CertificateStatus;
 import greencity.exceptions.BadRequestException;
 import greencity.exceptions.NotFoundException;
-import greencity.exceptions.certificate.CertificateIsNotActivated;
 import greencity.filters.CertificateFilterCriteria;
 import greencity.filters.CertificatePage;
 import greencity.repository.CertificateCriteriaRepo;
 import greencity.repository.CertificateRepository;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.Data;
 import org.modelmapper.ModelMapper;
@@ -58,8 +41,7 @@ public class CertificateServiceImpl implements CertificateService {
 
     @Override
     public void deleteCertificate(String code) {
-        Certificate certificate = certificateRepository.findById(code)
-            .orElseThrow(() -> new NotFoundException(CERTIFICATE_NOT_FOUND_BY_CODE + code));
+        Certificate certificate = findCertificateById(code);
         if (CertificateStatus.EXPIRED.equals(certificate.getCertificateStatus())
             || CertificateStatus.USED.equals(certificate.getCertificateStatus())) {
             throw new BadRequestException(ErrorMessage.CERTIFICATE_STATUS);
@@ -73,6 +55,27 @@ public class CertificateServiceImpl implements CertificateService {
         Page<Certificate> certificates =
             certificateCriteriaRepo.findAllWithFilter(certificatePage, certificateFilterCriteria);
         return getAllCertificatesTranslationDto(certificates);
+    }
+
+    @Override
+    public CertificateDto checkCertificate(String code, String userUuid) {
+        Certificate certificate = findCertificateById(code);
+
+        if (certificate.getCertificateStatus().equals(CertificateStatus.USED)
+            && !certificate.getOrder().getUser().getUuid().equals(userUuid)) {
+            return CertificateDto.builder()
+                .code(certificate.getCode())
+                .creationDate(certificate.getCreationDate())
+                .expirationDate(certificate.getExpirationDate())
+                .certificateStatus(certificate.getCertificateStatus().toString())
+                .build();
+        }
+        return modelMapper.map(certificate, CertificateDto.class);
+    }
+
+    private Certificate findCertificateById(String code) {
+        return certificateRepository.findById(code)
+            .orElseThrow(() -> new NotFoundException(CERTIFICATE_NOT_FOUND_BY_CODE + code));
     }
 
     private PageableDto<CertificateDtoForSearching> getAllCertificatesTranslationDto(Page<Certificate> pages) {
