+++ service/src/main/java/greencity/service/ubs/point/PointServiceImpl.java
@@ -1,0 +1,47 @@
+package greencity.service.ubs.point;
+
+import static greencity.constant.ErrorMessage.USER_WITH_CURRENT_ID_DOES_NOT_EXIST;
+import static java.util.Objects.nonNull;
+import greencity.dto.user.AllPointsUserDto;
+import greencity.dto.user.PointsForUbsUserDto;
+import greencity.entity.order.ChangeOfPoints;
+import greencity.entity.user.User;
+import greencity.exceptions.user.UserNotFoundException;
+import greencity.repository.UserRepository;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.modelmapper.ModelMapper;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class PointServiceImpl implements PointService {
+    private final UserRepository userRepository;
+    private final ModelMapper modelMapper;
+
+    @Override
+    public AllPointsUserDto findAllCurrentPointsForUser(String uuid) {
+        User currentUser = userRepository.findUserByUuid(uuid)
+            .orElseThrow(() -> new UserNotFoundException(USER_WITH_CURRENT_ID_DOES_NOT_EXIST));
+        Integer userBonuses = currentUser.getCurrentPoints();
+        if (userBonuses == null) {
+            userBonuses = 0;
+        }
+        List<ChangeOfPoints> changeOfPointsList = currentUser.getChangeOfPointsList();
+        List<PointsForUbsUserDto> bonusForUbsUser = new ArrayList<>();
+        if (nonNull(changeOfPointsList)) {
+            bonusForUbsUser = changeOfPointsList.stream()
+                .sorted(Comparator.comparing(ChangeOfPoints::getDate).reversed())
+                .map(m -> modelMapper.map(m, PointsForUbsUserDto.class))
+                .toList();
+        }
+        AllPointsUserDto allBonusesForUserDto = new AllPointsUserDto();
+        allBonusesForUserDto.setUserBonuses(userBonuses);
+        allBonusesForUserDto.setUbsUserBonuses(bonusForUbsUser);
+        return allBonusesForUserDto;
+    }
+}
